Verif que si Mahyco compile avec cuda, alors Arcane doit fournir la lib cuda_runtime.so

+++ bin/do_build_Mahyco.sh
@@ -127,6 +127,15 @@ else
   BASETMPDIR_NONREG="/tmp"
 fi
 
+
+ARCANE_CUDA_MODE=$(<${ARCANE_INSTALL_PATH}/${CCCOS}/${MHC_BUILD_TYPE}/lib/libarcane_accelerator_cuda_runtime.so)
+if [[ -z "${ARCANE_CUDA_MODE}" ]]
+then
+  echo "Lib libarcane_accelerator_cuda_runtime.so couldn't be found in ${ARCANE_INSTALL_PATH}. You will not be able to launch Mahyco with support of Nvidia GPUs."
+  echo "Please recompile Arcane with support of Nvidia GPUs."
+  exit 1
+fi
+
 rm -rf   build_${CCCOS}${MHC_MODE_SUFFIX_PART}/${MHC_BUILD_TYPE}
 mkdir -p build_${CCCOS}${MHC_MODE_SUFFIX_PART}/${MHC_BUILD_TYPE}
 cd       build_${CCCOS}${MHC_MODE_SUFFIX_PART}/${MHC_BUILD_TYPE}
